requirements.txt
streamlit
groq
python-dotenv
yt-dlp
openai